Complete move
