+++ src/bricks/transformers/ephemeralForm/EphemeralForm.test.tsx
@@ -15,12 +15,12 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { getFormDefinition } from "@/contentScript/messenger/api";
+import { getFormDefinition } from "@/contentScript/messenger/strict/api";
 import { render, screen } from "@testing-library/react";
 import React from "react";
 import EphemeralForm from "./EphemeralForm";
 
-jest.mock("@/contentScript/messenger/api");
+jest.mock("@/contentScript/messenger/strict/api");
 
 const getFormDefinitionMock = jest.mocked(getFormDefinition);
