+++ src/bricks/transformers/temporaryInfo/useTemporaryPanelDefinition.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { getPanelDefinition } from "@/contentScript/messenger/api";
+import { getPanelDefinition } from "@/contentScript/messenger/strict/api";
 import { type UUID } from "@/types/stringTypes";
 import { type TemporaryPanelEntry } from "@/types/sidebarTypes";
 import { type Target } from "@/types/messengerTypes";
