+++ src/pageEditor/sidebar/ActivatedModComponentListItem.tsx
@@ -30,11 +30,8 @@ import {
   ExtensionIcon,
   NotAvailableIcon,
 } from "@/pageEditor/sidebar/ExtensionIcons";
-import {
-  disableOverlay,
-  enableOverlay,
-  updateSidebar,
-} from "@/contentScript/messenger/api";
+import { disableOverlay, enableOverlay } from "@/contentScript/messenger/api";
+import { updateSidebar } from "@/contentScript/messenger/strict/api";
 import { openSidePanel } from "@/utils/sidePanelUtils";
 import { thisTab } from "@/pageEditor/utils";
 import cx from "classnames";
