+++ src/pageEditor/sidebar/DynamicModComponentListItem.tsx
@@ -27,11 +27,8 @@ import {
   UnsavedChangesIcon,
 } from "@/pageEditor/sidebar/ExtensionIcons";
 import { type UUID } from "@/types/stringTypes";
-import {
-  disableOverlay,
-  enableOverlay,
-  updateSidebar,
-} from "@/contentScript/messenger/api";
+import { disableOverlay, enableOverlay } from "@/contentScript/messenger/api";
+import { updateSidebar } from "@/contentScript/messenger/strict/api";
 import { openSidePanel } from "@/utils/sidePanelUtils";
 import { thisTab } from "@/pageEditor/utils";
 import cx from "classnames";
