+++ src/sidebar/ConnectedSidebar.tsx
@@ -38,10 +38,8 @@ import { selectIsSidebarEmpty } from "@/sidebar/sidebarSelectors";
 import DelayedRender from "@/components/DelayedRender";
 import DefaultPanel from "@/sidebar/DefaultPanel";
 import { MOD_LAUNCHER } from "@/sidebar/modLauncher/constants";
-import {
-  ensureExtensionPointsInstalled,
-  getReservedSidebarEntries,
-} from "@/contentScript/messenger/api";
+import { ensureExtensionPointsInstalled } from "@/contentScript/messenger/api";
+import { getReservedSidebarEntries } from "@/contentScript/messenger/strict/api";
 import { getConnectedTarget } from "@/sidebar/connectedTarget";
 import useAsyncEffect from "use-async-effect";
 import activateLinkClickHandler from "@/activation/activateLinkClickHandler";
