+++ src/sidebar/useHideEmptySidebar.ts
@@ -16,7 +16,7 @@
  */
 
 import useAsyncEffect from "use-async-effect";
-import { getReservedSidebarEntries } from "@/contentScript/messenger/api";
+import { getReservedSidebarEntries } from "@/contentScript/messenger/strict/api";
 import { closeSelf } from "@/sidebar/protocol";
 import { getConnectedTarget } from "@/sidebar/connectedTarget";
 import { useSelector } from "react-redux";
