+++ src/types/strictMessengerApi.ts
@@ -20,45 +20,3 @@
  * Make sure you add the methods you need to ./typeOnlyMessengerRegistration.ts
  * @see https://github.com/pixiebrix/pixiebrix-extension/issues/6526
  */
-
-import { getMethod, getNotifier } from "webext-messenger";
-
-export const showSidebar = getMethod("SHOW_SIDEBAR");
-export const hideSidebar = getMethod("HIDE_SIDEBAR");
-export const showMySidePanel = getMethod("SHOW_MY_SIDE_PANEL");
-
-// Content script
-export const getFormDefinition = getMethod("FORM_GET_DEFINITION");
-
-export const resolveForm = getMethod("FORM_RESOLVE");
-export const cancelForm = getNotifier("FORM_CANCEL");
-
-export const handleMenuAction = getMethod("HANDLE_MENU_ACTION");
-export const getReservedSidebarEntries = getMethod(
-  "GET_RESERVED_SIDEBAR_ENTRIES",
-);
-export const updateSidebar = getNotifier("UPDATE_SIDEBAR");
-export const sidebarWasLoaded = getNotifier("SIDEBAR_WAS_LOADED");
-export const reloadSidebar = getMethod("RELOAD_SIDEBAR");
-export const removeSidebars = getMethod("REMOVE_SIDEBARS");
-
-export const initRobot = getMethod("UIPATH_INIT");
-export const getProcesses = getMethod("UIPATH_GET_PROCESSES");
-
-export const closeWalkthroughModal = getNotifier("WALKTHROUGH_MODAL_CLOSE");
-export const showWalkthroughModal = getNotifier("WALKTHROUGH_MODAL_SHOW");
-
-export const getPanelDefinition = getMethod("PANEL_GET_DEFINITION");
-export const cancelTemporaryPanel = getNotifier("TEMPORARY_PANEL_CANCEL");
-export const closeTemporaryPanel = getNotifier("TEMPORARY_PANEL_CLOSE");
-export const resolveTemporaryPanel = getNotifier("TEMPORARY_PANEL_RESOLVE");
-
-export const getPageState = getMethod("GET_PAGE_STATE");
-export const setPageState = getMethod("SET_PAGE_STATE");
-
-export const notify = {
-  info: getNotifier("NOTIFY_INFO"),
-  // TODO: Automatically report from api.ts because of https://github.com/pixiebrix/pixiebrix-extension/blob/dce0d5cbb54d5fc1a61d720e43d17383a152df2e/src/background/contextMenus.ts#L92-L95
-  error: getNotifier("NOTIFY_ERROR"),
-  success: getNotifier("NOTIFY_SUCCESS"),
-};
